move internal/telemetry/trace => pkg/telemetry/trace
