+++ pkg/telemetry/trace/client.go
