+++ pkg/telemetry/trace/global.go
