+++ pkg/telemetry/trace/trace_export_test.go
